Move the exec-wrapper Run* functions into their own package (#521)
I'm about to make these more complex, and common shouldn't be such a
dumping ground of unrelated functionality.

There's nothing interesting in this PR, just moved code.
