+++ templates/common/git/git.go
@@ -27,6 +27,7 @@ import (
 	"github.com/Masterminds/semver/v3"
 
 	"github.com/abcxyz/abc/templates/common"
+	"github.com/abcxyz/abc/templates/common/run"
 )
 
 var sha = regexp.MustCompile("^[0-9a-f]{40}$")
@@ -40,17 +41,17 @@ var sha = regexp.MustCompile("^[0-9a-f]{40}$")
 // https://github.com/abcxyz/abc.git or git@github.com:abcxyz/abc.git .
 func Clone(ctx context.Context, remote, version, outDir string) error {
 	if sha.MatchString(version) {
-		_, _, err := common.Run(ctx, "git", "clone", remote, outDir)
+		_, _, err := run.Run(ctx, "git", "clone", remote, outDir)
 		if err != nil {
 			return err //nolint:wrapcheck
 		}
 
-		_, _, err = common.Run(ctx, "git", "-C", outDir, "reset", "--hard", version)
+		_, _, err = run.Run(ctx, "git", "-C", outDir, "reset", "--hard", version)
 		if err != nil {
 			return err //nolint:wrapcheck
 		}
 	} else {
-		_, _, err := common.Run(ctx, "git", "clone", "--depth", "1", "--branch", version, remote, outDir)
+		_, _, err := run.Run(ctx, "git", "clone", "--depth", "1", "--branch", version, remote, outDir)
 		if err != nil {
 			return err //nolint:wrapcheck
 		}
@@ -103,7 +104,7 @@ func findSymlinks(dir string) ([]string, error) {
 // "remote" may be any format accepted by git, such as
 // https://github.com/abcxyz/abc.git or git@github.com:abcxyz/abc.git .
 func RemoteTags(ctx context.Context, remote string) ([]string, error) {
-	stdout, _, err := common.Run(ctx, "git", "ls-remote", "--tags", remote)
+	stdout, _, err := run.Run(ctx, "git", "ls-remote", "--tags", remote)
 	if err != nil {
 		return nil, err //nolint:wrapcheck
 	}
@@ -173,7 +174,7 @@ func IsClean(ctx context.Context, dir string) (bool, error) {
 	// promises stable output, so it's good enough.
 	// https://stackoverflow.com/a/2658301
 	args := []string{"git", "-C", dir, "status", "--porcelain"}
-	stdout, _, err := common.Run(ctx, args...)
+	stdout, _, err := run.Run(ctx, args...)
 	if err != nil {
 		return false, err //nolint:wrapcheck
 	}
@@ -185,7 +186,7 @@ func IsClean(ctx context.Context, dir string) (bool, error) {
 // empty slice, this is not an error.
 func HeadTags(ctx context.Context, dir string) ([]string, error) {
 	args := []string{"git", "-C", dir, "for-each-ref", "--points-at", "HEAD", "refs/tags/*"}
-	stdout, _, err := common.Run(ctx, args...)
+	stdout, _, err := run.Run(ctx, args...)
 	if err != nil {
 		return nil, err //nolint:wrapcheck
 	}
@@ -213,7 +214,7 @@ func HeadTags(ctx context.Context, dir string) ([]string, error) {
 // workspace.
 func CurrentSHA(ctx context.Context, dir string) (string, error) {
 	args := []string{"git", "-C", dir, "rev-parse", "HEAD"}
-	stdout, _, err := common.Run(ctx, args...)
+	stdout, _, err := run.Run(ctx, args...)
 	if err != nil {
 		return "", err //nolint:wrapcheck
 	}
